Remove Estimator

+++ data_provider/shapestacks_provider.py
@@ -257,7 +257,7 @@ def shapestacks_input_fn(
   # parse data from files and apply pre-processing
   filenames = tf.constant(filenames)
   labels = tf.constant(labels)
-  images = tf.map_fn(_parse_record,filenames)
+  images = tf.map_fn(_parse_record,filenames,dtype=tf.float32)
   print(images)
   if augment != [] and mode == 'train':
     dataset = dataset.map(lambda feature, label: _augment(feature, label, augment))
